+++ api/cpp/cbindgen.rs
@@ -117,6 +117,7 @@ fn gen_corelib(
         "ImageFit",
         "FillRule",
         "MouseCursor",
+        "InputType",
         "StandardButtonKind",
         "DialogButtonRole",
         "PointerEventKind",
@@ -226,6 +227,7 @@ fn gen_corelib(
                 "slint_image_path",
                 "SharedPixelBuffer",
                 "SharedImageBuffer",
+                "StaticTextures",
             ],
             vec!["Color"],
             "slint_image_internal.h",
@@ -433,13 +435,15 @@ fn gen_backend_qt(
         "NativeComboBoxPopup",
         "NativeTabWidget",
         "NativeTab",
+        "NativeStyleMetrics",
     ];
 
     config.export.include = items.iter().map(|x| x.to_string()).collect();
 
     config.export.body.insert(
         "NativeStyleMetrics".to_owned(),
-        "    inline NativeStyleMetrics(); inline ~NativeStyleMetrics();".to_owned(),
+        "    inline explicit NativeStyleMetrics(void* = nullptr); inline ~NativeStyleMetrics();"
+            .to_owned(),
     );
 
     let mut crate_dir = root_dir.to_owned();
@@ -455,6 +459,31 @@ fn gen_backend_qt(
         .generate()
         .context("Unable to generate bindings for slint_qt_internal.h")?
         .write_to_file(include_dir.join("slint_qt_internal.h"));
+
+    Ok(())
+}
+
+fn gen_backend_selector(
+    root_dir: &Path,
+    include_dir: &Path,
+    dependencies: &mut Vec<PathBuf>,
+) -> anyhow::Result<()> {
+    let mut config = default_config();
+
+    config.export.include.clear();
+
+    let mut crate_dir = root_dir.to_owned();
+    crate_dir.extend(["internal", "backends", "selector"].iter());
+
+    ensure_cargo_rerun_for_crate(&crate_dir, dependencies)?;
+
+    cbindgen::Builder::new()
+        .with_config(config)
+        .with_crate(crate_dir)
+        .with_include("slint_qt_internal.h")
+        .generate()
+        .context("Unable to generate bindings for slint_selector_internal.h")?
+        .write_to_file(include_dir.join("slint_selector_internal.h"));
 
     Ok(())
 }
@@ -560,6 +589,7 @@ pub fn gen_all(root_dir: &Path, include_dir: &Path) -> anyhow::Result<Vec<PathBu
     let mut deps = Vec::new();
     gen_corelib(root_dir, include_dir, &mut deps)?;
     gen_backend_qt(root_dir, include_dir, &mut deps)?;
+    gen_backend_selector(root_dir, include_dir, &mut deps)?;
     gen_backend(root_dir, include_dir, &mut deps)?;
     gen_interpreter(root_dir, include_dir, &mut deps)?;
     Ok(deps)