+++ internal/backends/gl/fonts/fontconfig.rs
@@ -3,6 +3,7 @@
 
 use fontconfig::fontconfig;
 
+// This is duplicated in the slint-compiler's glyph embedding code
 pub fn find_families(requested_family: &str) -> Vec<String> {
     unsafe {
         let config = fontconfig::FcInitLoadConfigAndFonts();