+++ internal/backends/gl/lib.rs
@@ -4,6 +4,8 @@
 #![doc = include_str!("README.md")]
 #![doc(html_logo_url = "https://slint-ui.com/logo/slint-logo-square-light.svg")]
 
+extern crate alloc;
+
 use std::cell::RefCell;
 use std::pin::Pin;
 use std::rc::Rc;
@@ -14,9 +16,10 @@ use i_slint_core::graphics::{
     Brush, Color, Image, ImageInner, IntRect, IntSize, Point, Rect, RenderingCache, Size,
 };
 use i_slint_core::item_rendering::{CachedRenderingData, ItemRenderer};
-use i_slint_core::items::{FillRule, ImageFit, ImageRendering};
+use i_slint_core::items::{FillRule, ImageFit, ImageRendering, InputType};
 use i_slint_core::properties::Property;
 use i_slint_core::window::{Window, WindowRc};
+use i_slint_core::SharedString;
 
 mod glwindow;
 use glwindow::*;
@@ -28,6 +31,8 @@ mod svg;
 use images::*;
 
 mod fonts;
+
+mod stylemetrics;
 
 type Canvas = femtovg::Canvas<femtovg::renderer::OpenGl>;
 type CanvasRc = Rc<RefCell<Canvas>>;
@@ -273,12 +278,20 @@ impl ItemRenderer for GLItemRenderer {
             None => return,
         };
 
-        let (min_select, max_select) = text_input.selection_anchor_and_cursor();
+        let (mut min_select, mut max_select) = text_input.selection_anchor_and_cursor();
         let cursor_pos = text_input.cursor_position();
         let cursor_visible = cursor_pos >= 0 && text_input.cursor_visible() && text_input.enabled();
+        let mut cursor_pos = cursor_pos as usize;
         let mut canvas = self.canvas.borrow_mut();
         let font_height = canvas.measure_font(paint).unwrap().height();
-        let text = text_input.text();
+        let mut text = text_input.text();
+
+        if let InputType::password = text_input.input_type() {
+            min_select = text[..min_select].chars().count() * PASSWORD_CHARACTER.len();
+            max_select = text[..max_select].chars().count() * PASSWORD_CHARACTER.len();
+            cursor_pos = text[..cursor_pos].chars().count() * PASSWORD_CHARACTER.len();
+            text = SharedString::from(PASSWORD_CHARACTER.repeat(text.chars().count()));
+        };
 
         let mut cursor_point: Option<Point> = None;
 
@@ -365,7 +378,6 @@ impl ItemRenderer for GLItemRenderer {
                     // no selection on this line
                     canvas.fill_text(pos.x, pos.y, to_draw.trim_end(), paint).unwrap();
                 };
-                let cursor_pos = cursor_pos as usize;
                 if cursor_visible
                     && (range.contains(&cursor_pos)
                         || (cursor_pos == range.end && cursor_pos == text.len()))
@@ -1152,10 +1164,14 @@ pub fn create_gl_window_with_canvas_id(canvas_id: String) -> Rc<Window> {
 pub fn use_modules() {}
 
 pub type NativeWidgets = ();
-pub type NativeGlobals = ();
-pub mod native_widgets {}
+pub type NativeGlobals = (stylemetrics::NativeStyleMetrics, ());
+pub mod native_widgets {
+    pub use super::stylemetrics::NativeStyleMetrics;
+}
 pub const HAS_NATIVE_STYLE: bool = false;
-pub const IS_AVAILABLE: bool = true;
+
+pub use stylemetrics::native_style_metrics_deinit;
+pub use stylemetrics::native_style_metrics_init;
 
 // TODO: We can't connect to the wayland clipboard yet because
 // it requires an external connection.