+++ internal/backends/qt/qt_widgets/lineedit.rs
@@ -1,6 +1,9 @@
 // Copyright © SixtyFPS GmbH <info@slint-ui.com>
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-commercial
 
+use i_slint_core::input::FocusEventResult;
+use i_slint_core::items::InputType;
+
 use super::*;
 
 #[repr(C)]
@@ -18,6 +21,7 @@ pub struct NativeLineEdit {
     pub native_padding_bottom: Property<f32>,
     pub has_focus: Property<bool>,
     pub enabled: Property<bool>,
+    pub input_type: Property<InputType>,
 }
 
 impl Item for NativeLineEdit {
@@ -102,7 +106,9 @@ impl Item for NativeLineEdit {
         KeyEventResult::EventIgnored
     }
 
-    fn focus_event(self: Pin<&Self>, _: &FocusEvent, _window: &WindowRc) {}
+    fn focus_event(self: Pin<&Self>, _: &FocusEvent, _window: &WindowRc) -> FocusEventResult {
+        FocusEventResult::FocusIgnored
+    }
 
     fn_render! { this dpr size painter widget initial_state =>
         let has_focus: bool = this.has_focus();